refactor(mf2): Collect all type guards into src/data-model/type-guards.ts
