+++ packages/mf2-messageformat/src/data-model/expression/literal.ts
@@ -16,15 +16,6 @@ export interface Literal {
   value: string;
 }
 
-/**
- * Type guard for {@link Literal} pattern elements
- *
- * @beta
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const isLiteral = (part: any): part is Literal =>
-  !!part && typeof part === 'object' && part.type === 'literal';
-
 export function resolveLiteral(ctx: Context, lit: Literal) {
   const msgCtx = new MessageFunctionContext(ctx, `|${lit.value}|`);
   return ctx.functions.string(msgCtx, {}, lit.value);
